requirements.txt
PyQt6==6.6.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.0
alembic==1.12.1
python-dateutil==2.8.2